chkout

+++ travelers_choice/lib/controllers/profile_controller.dart
@@ -13,7 +13,8 @@ class ProfileController extends FxController {
   bool notification = true, offlineReading = false;
   late User user;
   String? name, email;
-  double? balanceamount;
+  // double? balanceamount;
+  int? balanceamount;
 
   @override
   initState() {
